+++ www/src/components/layout.js
@@ -1,7 +1,7 @@
 import React from "react"
 import Modal from "react-modal"
-import Helmet from "react-helmet"
 import { SkipNavLink } from "@reach/skip-nav"
+import { OutboundLink } from "gatsby-plugin-google-analytics"
 import MdClose from "react-icons/lib/md/close"
 import { navigate, PageRenderer } from "gatsby"
 import presets, { colors } from "../utils/presets"
@@ -9,6 +9,7 @@ import Banner from "../components/banner"
 import Navigation from "../components/navigation"
 import MobileNavigation from "../components/navigation-mobile"
 import PageWithSidebar from "../components/page-with-sidebar"
+import SiteMetadata from "../components/site-metadata"
 
 import mousetrap from "mousetrap"
 
@@ -147,40 +148,22 @@ class DefaultLayout extends React.Component {
 
     return (
       <div className={isHomepage ? `is-homepage` : ``}>
-        <Helmet defaultTitle={`GatsbyJS`} titleTemplate={`%s | GatsbyJS`}>
-          <meta name="twitter:site" content="@gatsbyjs" />
-          <meta name="og:type" content="website" />
-          <meta name="og:site_name" content="GatsbyJS" />
-          <link
-            rel="canonical"
-            href={`https://gatsbyjs.org${this.props.location.pathname}`}
-          />
-          <html lang="en" />
-        </Helmet>
+        <SiteMetadata pathname={this.props.location.pathname} />
         <SkipNavLink css={styles.skipLink}>Skip to main content</SkipNavLink>
         <Banner background={isHomepage ? `#402060` : false}>
-          These are the docs for v2.
-          {` `}
-          <a
-            href="https://v1.gatsbyjs.org/"
+          {/* !!! If you change the children of Banner remember to do the same in layout/layout-with-heading.js */}
+          <OutboundLink
+            href="https://www.gatsbyjs.com/content-mesh-contentful"
             css={{
               color: `#fff`,
+              "&:hover": {
+                color: `#fff`,
+              },
             }}
           >
-            View the v1 docs
-            <span
-              css={{
-                display: `none`,
-                [presets.Mobile]: {
-                  display: `inline`,
-                },
-              }}
-            >
-              {` `}
-              instead
-            </span>
-          </a>
-          .
+            Watch
+          </OutboundLink>
+          {`: “Rise of the Content Mesh: Webcast with Contentful and Gatsby”.`}
         </Banner>
         <Navigation pathname={this.props.location.pathname} />
         <div
@@ -193,6 +176,8 @@ class DefaultLayout extends React.Component {
                 ? presets.bannerHeight
                 : `calc(${presets.bannerHeight} + ${presets.headerHeight})`,
             },
+            paddingLeft: `env(safe-area-inset-left)`,
+            paddingRight: `env(safe-area-inset-right)`,
           }}
         >
           <PageWithSidebar
@@ -211,28 +196,26 @@ class DefaultLayout extends React.Component {
 
 const styles = {
   skipLink: {
-    "[data-reach-skip-link]": {
-      border: `0`,
-      clip: `rect(0 0 0 0)`,
-      height: `1px`,
-      width: `1px`,
-      margin: `-1px`,
-      padding: `0`,
-      overflow: `hidden`,
-      position: `absolute`,
-      zIndex: `100`,
-      fontSize: `0.85rem`,
-      ":focus": {
-        padding: `0.9rem`,
-        position: `fixed`,
-        top: `10px`,
-        left: `10px`,
-        background: `white`,
-        textDecoration: `none`,
-        width: `auto`,
-        height: `auto`,
-        clip: `auto`,
-      },
+    border: `0`,
+    clip: `rect(0 0 0 0)`,
+    height: 1,
+    width: 1,
+    margin: -1,
+    padding: 0,
+    overflow: `hidden`,
+    position: `absolute`,
+    zIndex: 100,
+    fontSize: `0.85rem`,
+    ":focus": {
+      padding: `0.9rem`,
+      position: `fixed`,
+      top: 10,
+      left: 10,
+      background: `white`,
+      textDecoration: `none`,
+      width: `auto`,
+      height: `auto`,
+      clip: `auto`,
     },
   },
 }